Split EcoEnergyDbService class using partial, to divide functionality.
